add some new stuff
