--- cypress/support/customCommands.d.ts
@@ -1,5 +1,0 @@
-declare namespace Cypress {
-    interface Chainable<Subject> {
-      clickToMenuByName(element: Cypress.Chainable): Chainable<any>
-  }
-}
