+++ cypress/support/other_commands.js
@@ -1,0 +1,22 @@
+Cypress.Commands.add("getPrimaryBtn", { prevSubject: "optional" }, (subj) => {
+    if (subj) {
+      cy.wrap(subj).find("[status='primary']");
+    } else {
+      cy.get("[status='primary']");
+    }
+  });
+  
+  Cypress.Commands.overwrite("type", (originalFn, element, text, options) => {
+    if (options && options.sensitive) {
+      // turn off original log
+      options.log = false;
+      // create our own log with masked message
+      Cypress.log({
+        $el: element,
+        name: "type",
+        message: "*".repeat(text.length),
+      });
+    }
+  
+    return originalFn(element, text, options);
+  });
